+++ src/_pytest/deprecated.py
@@ -46,11 +46,6 @@ YIELD_FIXTURE = PytestDeprecationWarning(
     "Use @pytest.fixture instead; they are the same."
 )
 
-WARNING_CMDLINE_PREPARSE_HOOK = PytestRemovedIn8Warning(
-    "The pytest_cmdline_preparse hook is deprecated and will be removed in a future release. \n"
-    "Please use pytest_load_initial_conftests hook instead."
-)
-
 STRICT_OPTION = PytestRemovedIn8Warning(
     "The --strict option is deprecated, use --strict-markers instead."
 )